Fix for #2127: Update Live555 library - ServerMediaSession update caused hangs on seeking and channel switching

+++ DirectShowFilters/LiveMedia555/liveMedia/ServerMediaSession.hh
@@ -135,7 +135,7 @@ public:
 			   unsigned short& rtpSeqNum,
 			   unsigned& rtpTimestamp) = 0;
   virtual void pauseStream(unsigned clientSessionId, void* streamToken);
-  virtual void seekStream(unsigned clientSessionId, void* streamToken, double seekNPT);
+  virtual void seekStream(unsigned clientSessionId, void* streamToken, float seekNPT);
   virtual void setStreamScale(unsigned clientSessionId, void* streamToken, float scale);
   virtual void deleteStream(unsigned clientSessionId, void*& streamToken);
